requirements.txt
streamlit>=1.31.0
tensorflow>=2.15.0,<2.16.0
numpy>=1.24.3,<2.0.0
Pillow>=10.2.0
google-generativeai>=0.3.2
protobuf>=3.20.3,<4.0.0
requests>=2.31.0
